[ALPS04016653] DISP: fix display build error in 32bit
[Detail]
fix display linker error in kernel 32bit project

MTK-Commit-Id: 9894dcebc164dd658657764810411575531ad7ab

CR-Id: ALPS04016653
Feature: Display Driver (DSI Interface)
(cherry-pick from 3e0aaf43613d46eb11bafd19090dd80b1e0d1d41)

+++ drivers/misc/mediatek/video/mt6771/videox/debug.h
@@ -51,7 +51,7 @@ extern int hrt_show_flag;
 extern int layer_en_num;
 extern int layer_en_num_flag;
 extern uint64_t layer_size_high;
-extern int layer_size_low;
+extern uint64_t layer_size_low;
 extern int layer_size_flag;
 
 /* path mode */
